projection

+++ src/main/java/com/devsuperior/dsmeta/projections/SellerSumProjection.java
@@ -2,5 +2,5 @@ package com.devsuperior.dsmeta.projections;
 
 public interface SellerSumProjection {
     String getName();
-    Long getSum();
+    Double getSum();
 }
